requirements.txt
Django
South
django-tastypie
-e git+https://github.com/zoranzaric/django-gal.git#egg=django-gal
